Refactor film rendering to be inside a controller
